Refactor range layouts
