+++ console/src/range/external.ts
@@ -8,12 +8,12 @@
 // included in the file licenses/APL.txt.
 
 import { type Layout } from "@/layout";
-import { Create, CREATE_LAYOUT_TYPE } from "@/range/CreateLayout";
-import { OVERVIEW_LAYOUT } from "@/range/overview/layout";
+import { Create, CREATE_LAYOUT_TYPE } from "@/range/Create";
+import { OVERVIEW_LAYOUT_TYPE } from "@/range/overview/layout";
 import { Overview } from "@/range/overview/Overview";
 
 export * from "@/range/ContextMenu";
-export * from "@/range/CreateLayout";
+export * from "@/range/Create";
 export * from "@/range/overview/layout";
 export * from "@/range/overview/Overview";
 export * from "@/range/Select";
@@ -24,5 +24,5 @@ export * from "@/range/Toolbar";
 
 export const LAYOUTS: Record<string, Layout.Renderer> = {
   [CREATE_LAYOUT_TYPE]: Create,
-  [OVERVIEW_LAYOUT.type]: Overview,
+  [OVERVIEW_LAYOUT_TYPE]: Overview,
 };
